requirements.txt
psutil
docker
netifaces
luma.oled
Pillow
